Updated HBaseMetadata to use a common path for performing mutations

+++ HBaseAdapter/src/main/java/com/nearinfinity/honeycomb/hbase/HBaseMetadata.java
@@ -97,8 +97,7 @@ public class HBaseMetadata {
         puts.add(putColumnIds(tableId, schema.getColumns()));
         puts.add(putTableSchema(tableId, schema));
 
-        hTable.put(puts);
-        hTable.flushCommits();
+        performMutations(ImmutableList.<Delete>of(), puts);
     }
 
 
@@ -122,8 +121,7 @@ public class HBaseMetadata {
         deletes.add(deleteAutoIncCounter(tableId));
         deletes.add(deleteTableSchema(tableId));
 
-        hTable.delete(deletes);
-        hTable.flushCommits();
+        performMutations(deletes, ImmutableList.<Put>of());
     }
 
     public void updateSchema(long tableId, TableSchema oldSchema, TableSchema newSchema)
@@ -203,8 +201,8 @@ public class HBaseMetadata {
     }
 
     public void truncateAutoInc(long tableId) throws IOException {
-        hTable.delete(deleteAutoIncCounter(tableId));
-        hTable.flushCommits();
+        performMutations(ImmutableList.<Delete>of(deleteAutoIncCounter(tableId)),
+                ImmutableList.<Put>of());
     }
 
     public long getRowCount(long tableId) throws IOException {
@@ -216,8 +214,8 @@ public class HBaseMetadata {
     }
 
     public void truncateRowCount(long tableId) throws IOException {
-        hTable.delete(deleteRowsCounter(tableId));
-        hTable.flushCommits();
+        performMutations(ImmutableList.<Delete>of(deleteRowsCounter(tableId)),
+                ImmutableList.<Put>of());
     }
 
     private long getCounter(byte[] row, byte[] identifier) throws IOException {
@@ -303,9 +301,29 @@ public class HBaseMetadata {
         return VarEncoder.decodeULong(id);
     }
 
+    /**
+     * Performs the operations necessary to commit the specified mutations to the underlying data store.
+     * If no operations are specified then no mutations occur.
+     * @param deletes A list of {@link Delete} operations to execute, not null
+     * @param puts A list of  {@link Put} operations to execute, not null
+     * @throws IOException Thrown on mutation commit failure
+     */
     private void performMutations(final List<Delete> deletes, final List<Put> puts) throws IOException {
-        hTable.delete(deletes);
-        hTable.put(puts);
-        hTable.flushCommits();
+        Preconditions.checkNotNull(deletes, "The delete mutations container is invalid");
+        Preconditions.checkNotNull(puts, "The put mutations container is invalid");
+        Preconditions.checkArgument(!deletes.isEmpty() || !puts.isEmpty(), "At least one mutation operation must be specified");
+
+        if( !deletes.isEmpty() ) {
+            hTable.delete(deletes);
+        }
+
+        if( !puts.isEmpty() ) {
+            hTable.put(puts);
+        }
+
+        // Only flush if the table is not configured to auto flush
+        if( !hTable.isAutoFlush() ) {
+            hTable.flushCommits();
+        }
     }
 }
